let's try cleaning up the swift.yml (#36)
Streamline the swift.yml file a little bit and add some static analysis. The old file essentially did nothing. It generated a new project, and built/ran the tests in this new project, without ever hitting the app code. Now it'll build the app code, analyze it, and test it.