Incorporate Quick and Nimble into our project and run pod update (updated mockingbird from 0.10.0 to 0.11.0) (#40)
Add Quick and Nimble into the project for better unit tests. See https://www.raywenderlich.com/135-behavior-driven-testing-tutorial-for-ios-with-quick-nimble for a good tutorial on what they are and how to use them

For some reason, a simple pod install doesn't work, so we are forced to add a pod deintegrate into the build step. There is an issue reported in the Nimble library and I also commented on the issue (Quick/Nimble#702 (comment)). Until we get more guidance on how to fix this CwlCatchException.h not found error, we're stuck deintegrating and re-installing all pods on every build. :(